(Widescreen2D)
Work-in-progress version of rendering the HUD non-stretched. Currently a bug where parts of the ammo pips aren't rendered correctly. Also added the r_stretch2D cvar for reverting back to the original behaviour of always stretching 2D elements.

Also made it so aspectRatio value is always automatically calculated, even if a custom resolution is defined.