(HigherEntityMax)
Fixed the issue I mention in the previous commit. I thought it was related to graphics rendering, but it was related to entity data being transmitted from server to client. The server keeps sending "snapshots" containing information about the current entities to the client (I think it tries to only send data about entities the player can see), but by default, this is limited to 256 (there are several limits that needs to be udpated, the main one being in sv_init.cpp). Once exceeded, the game gives up on sending info about all entities and it also doesn't do a "delta" snapshot (which I think is merging old and new entity data), which I think ends up deleting all outdated entity data.

The end result was that the client would stop seeing random entities once the limit was reached. But with the limit is increased, it seems to be fine.

I increased all the limits by quadrupling them (it matches the quadrupling of the game entity max). I also increased the effect count to match the max entity count.

I should note that increasing entity limit has made older savegame incompatible.