(Widescreen2D)
Fixed the rendering of the Disruptor scope. This required one small change to one of the assets. Currently, we need the main scope texture to use clamp as wrapping mode so one of the shader files need that modification (alternatively, we could change the code so it draws 2 extra quads that are stretched from the first and last texels).